requirements.txt
byol_pytorch==0.5.7
torch==1.7.1
pandarallel==1.5.2
wandb==0.10.26
plotly==4.14.3
torchmetrics==0.5.0
numpy==1.19.2
more_itertools==8.7.0
torchsummary==1.5.1
timm==0.4.5
pandas==1.1.4
stuf==0.9.16
torchdata==0.2.0
omegaconf==2.1.0.dev24
munch==2.5.0
pytorch_lightning==1.4.0
scikit_image==0.17.2
seaborn==0.11.1
python_box==5.3.0
hydra_core==1.1.0.dev5
lightning_bolts==0.3.4
ray==1.5.2
tqdm==4.49.0
matplotlib==3.2.2
modin==0.10.1
torchvision==0.8.2
plumbum==1.7.0
boltons==20.2.1
opencv_python==4.5.1.48
dask==2021.5.1
auto_mix_prep==0.2.0
dataclasses==0.8
fastai==2.5.2
hydra==2.5
ipdb==0.13.9
ipython==7.27.0
Pillow==8.3.1
python-dotenv==0.19.0
rich==10.9.0
scikit_learn==0.24.2
skimage==0.0
